refactor(api): split DashboardAuthHandler into 5 files by auth flow
Split dashboard_auth_handler.go (1570 lines) into focused files:
- dashboard_auth_handler.go: struct, constructor, core auth, middleware
- dashboard_auth_handler_mfa.go: TOTP setup/enable/disable
- dashboard_auth_handler_password.go: password reset flow
- dashboard_auth_handler_oauth.go: OAuth SSO login/callback/providers
- dashboard_auth_handler_saml.go: SAML login/ACS callback
